feat: separate server
